+++ release_notes.md
@@ -8,15 +8,15 @@
 
 ### 🔧 Compatibility Notes
 
-- 
+- [Connectivity] Destination objects with different header providers are now considered not equal.
 
 ### ✨ New Functionality
 
-- [OpenAPI] Added experimental support for generating clients from OpenAPI 3.1 specifications
+- 
 
 ### 📈 Improvements
 
-- 
+- [Connectivity] Fixed `Connection pool shut down` edge-case when closing a client with a destination that has a custom header provider.
 
 ### 🐛 Fixed Issues